Separei o front que a Aline fez em páginas e componentes. Criei uma pasta components e outra pages no src.
