--- client/src/app.css
@@ -1,143 +1,0 @@
-/* Estilos globais */
-body {
-  margin: 0;
-  font-family: 'Arial', sans-serif;
-  background-color: #f7f7f7;
-  color: #333;
-}
-
-/* Cabeçalho */
-header {
-  background-color: #4f46e5; /* Indigo-600 */
-  color: white;
-  padding: 1rem;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-}
-
-/* Imagem com tamanho fixo */
-.fixed-img {
-  width: 150px;
-  height: 150px;
-  object-fit: cover; /* Ajusta a imagem para cobrir o espaço sem distorção */
-  border-radius: 0.5rem; /* Cantos levemente arredondados */
-  margin: auto;
-  display: block;
-}
-
-header h1 {
-  font-size: 1.75rem;
-  font-weight: bold;
-}
-
-header nav ul {
-  list-style: none;
-  padding: 0;
-  margin: 0;
-  display: flex;
-  gap: 1rem;
-}
-
-header nav ul li a {
-  text-decoration: none;
-  color: white;
-  transition: color 0.3s ease;
-}
-
-header nav ul li a:hover {
-  color: #c3dafe; /* Indigo-200 */
-}
-
-/* Conteúdo principal */
-main {
-  flex-grow: 1;
-  padding: 2rem;
-  max-width: 1200px;
-  margin: auto;
-}
-
-section {
-  margin-bottom: 2rem;
-}
-
-section h2 {
-  font-size: 1.5rem;
-  font-weight: 600;
-  margin-bottom: 1rem;
-}
-
-section p {
-  font-size: 1rem;
-  color: #555;
-}
-
-.grid {
-  display: grid;
-  gap: 1rem;
-}
-
-.grid-cols-1 {
-  grid-template-columns: 1fr;
-}
-
-.grid-cols-2 {
-  grid-template-columns: repeat(2, 1fr);
-}
-
-.grid-cols-3 {
-  grid-template-columns: repeat(3, 1fr);
-}
-
-.grid-cols-4 {
-  grid-template-columns: repeat(4, 1fr);
-}
-
-/* Cartões */
-.card {
-  background-color: white;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-  border-radius: 0.5rem;
-  padding: 1rem;
-}
-
-.card h3 {
-  font-size: 1rem;
-  font-weight: bold;
-  margin-bottom: 0.5rem;
-}
-
-.card p {
-  font-size: 0.875rem;
-  color: #666;
-}
-
-.card span {
-  display: inline-block;
-  margin-top: 0.5rem;
-  font-size: 0.75rem;
-  font-weight: 600;
-  color: #4f46e5; /* Indigo-600 */
-}
-
-/* Botões */
-button {
-  background-color: #4f46e5; /* Indigo-600 */
-  color: white;
-  border: none;
-  padding: 0.5rem 1rem;
-  border-radius: 0.25rem;
-  cursor: pointer;
-  transition: background-color 0.3s ease;
-}
-
-button:hover {
-  background-color: #4338ca; /* Indigo-700 */
-}
-
-/* Rodapé */
-footer {
-  background-color: #333;
-  color: white;
-  text-align: center;
-  padding: 1rem;
-  font-size: 0.875rem;
-}
