+++ client/src/components/Footer/Footer.jsx
@@ -1,0 +1,13 @@
+import React from 'react'
+
+const Footer = () => {
+  return (
+    <div>
+      <footer className="bg-gray-800 text-white text-center p-4 text-sm">
+        <p>© 2024 Literar. Todos os direitos reservados.</p>
+      </footer>
+    </div>
+  )
+}
+
+export default Footer
